Refactor init.d control scripts
The goal of this change is to factor out and remove duplication in
instance specific init.d control scripts. This change adds the control
script to the installer and then updates the init.d script to invoke it.
The change should go even further and merge tomcat-ad.sh and data-match
scripts into one as they fulfill almost the same function. This could be
done once we validate this refactoring works.